etnaviv: fix buffer cache
We also need to use the aligned size when actually allocating a
new GEM object, otherwise the cache may return buffers that are
too small to hold the required amount of data.

Also don't double free the device reference when going through
the buffer cache free path. This fixes a memory corruption.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>
Acked-by: Christian Gmeiner <christian.gmeiner@gmail.com>